move variable encoding to expression module
