+++ packages/react-sdk/src/embed-template.ts
@@ -12,26 +12,7 @@ import {
 } from "@webstudio-is/project-build";
 import { StyleValue, type StyleProperty } from "@webstudio-is/css-data";
 import type { Simplify } from "type-fest";
-import { validateExpression } from "./expression";
-
-const dataSourceVariablePrefix = "$ws$dataSource$";
-
-// data source id is generated with nanoid which has "-" in alphabeta
-// here "-" is encoded with "__DASH__' in variable name
-// https://github.com/ai/nanoid/blob/047686abad8f15aff05f3a2eeedb7c98b6847392/url-alphabet/index.js
-
-export const encodeDataSourceVariable = (id: DataSource["id"]) => {
-  const encoded = id.replaceAll("-", "__DASH__");
-  return `${dataSourceVariablePrefix}${encoded}`;
-};
-
-export const decodeDataSourceVariable = (name: string) => {
-  if (name.startsWith(dataSourceVariablePrefix)) {
-    const encoded = name.slice(dataSourceVariablePrefix.length);
-    return encoded.replaceAll("__DASH__", "-");
-  }
-  return;
-};
+import { encodeDataSourceVariable, validateExpression } from "./expression";
 
 const EmbedTemplateText = z.object({
   type: z.literal("text"),
