+++ packages/react-sdk/src/index.ts
@@ -20,4 +20,9 @@ export {
   getInstanceIdFromComponentProps,
 } from "./props";
 export { type Params, ReactSdkContext } from "./context";
-export { validateExpression, executeExpressions } from "./expression";
+export {
+  validateExpression,
+  executeExpressions,
+  encodeDataSourceVariable,
+  decodeDataSourceVariable,
+} from "./expression";
